+++ DearPyGui/src/core/mvCallbackRegistry.cpp
@@ -177,21 +177,6 @@ namespace Marvel {
 		if (PyErr_Occurred())
 			PyErr_Print();
 
-		PyObject* intermediateResult = nullptr;
-		if (PyCallable_Check(user_data))
-		{
-			intermediateResult = PyObject_CallObject(user_data, nullptr);
-			// check if call succeeded
-			if (intermediateResult == nullptr)
-			{
-				PyErr_Print();
-				intermediateResult = user_data;
-			}
-
-		}
-		else
-			intermediateResult = user_data;
-
 		if (PyErr_Occurred())
 			PyErr_Print();
 
@@ -209,7 +194,7 @@ namespace Marvel {
 					mvPyObject pArgs(PyTuple_New(count));
 					PyTuple_SetItem(pArgs, 0, ToPyUUID(sender));
 					PyTuple_SetItem(pArgs, 1, app_data); // steals data, so don't deref
-					PyTuple_SetItem(pArgs, 2, intermediateResult); // steals data, so don't deref
+					PyTuple_SetItem(pArgs, 2, user_data); // steals data, so don't deref
 					
 					for (int i = 3; i < count; i++)
 						PyTuple_SetItem(pArgs, i, GetPyNone());
@@ -226,7 +211,7 @@ namespace Marvel {
 					mvPyObject pArgs(PyTuple_New(3));
 					PyTuple_SetItem(pArgs, 0, ToPyUUID(sender));
 					PyTuple_SetItem(pArgs, 1, app_data); // steals data, so don't deref
-					PyTuple_SetItem(pArgs, 2, intermediateResult); // steals data, so don't deref
+					PyTuple_SetItem(pArgs, 2, user_data); // steals data, so don't deref
 
 					mvPyObject result(PyObject_CallObject(callable, pArgs));